backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
async-timeout==4.0.3
base58==2.1.1
base58check==1.0.2
bech32==1.2.0
bech32m==1.0.0
bitcoin==1.1.42
bitcoin-utils==0.6.5
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
dnspython==2.5.0
ecdsa==0.17.0
fastapi==0.109.2
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
hdwallet==2.2.1
httptools==0.6.1
idna==3.6
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mnemonic==0.21
mpmath==1.3.0
Naked==0.1.32
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pydantic==2.6.1
pydantic-mongo==2.1.0
pydantic_core==2.16.2
pymongo==4.6.1
python-bitcoinrpc==1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
